+++ orderManager/src/main/java/com/challenge/orderManager/repositories/ProductRepositoryQueries.java
@@ -1,10 +1,11 @@
 package com.challenge.orderManager.repositories;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.challenge.orderManager.entities.Product;
 
 public interface ProductRepositoryQueries {	
-	Product getProduct(int productId);
+	Product getProduct(long productId);
 	List<Product> getProductList();
 }
